+++ src/N98/Magento/Command/Developer/Translate/InlineShopCommand.php
@@ -25,4 +25,15 @@ class InlineShopCommand extends AbstractMagentoStoreConfigCommand
      * @var string
      */
     protected $commandDescription = 'Toggle inline translation tool for shop';
+
+    /**
+     * If required, handle the output and possible change of the developer IP restrictions
+     *
+     * @param \Mage_Core_Model_Store $store
+     * @param bool $disabled
+     */
+    protected function _afterSave(\Mage_Core_Model_Store $store, $disabled)
+    {
+        $this->detectAskAndSetDeveloperIp($store, $disabled);
+    }
 }