+++ src/N98/Magento/Command/System/Setup/RemoveCommand.php
@@ -3,14 +3,15 @@
 namespace N98\Magento\Command\System\Setup;
 
 use InvalidArgumentException;
+use Mage_Core_Model_Resource;
+use RuntimeException;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
-use N98\Util\Console\Helper\Table\Renderer\RendererFactory;
 
 /**
  * Class RemoveCommand
+ *
  * @package N98\Magento\Command\System\Setup
  * @author Aydin Hassan <aydin@hotmail.co.uk>
  */
@@ -29,7 +30,7 @@ class RemoveCommand extends AbstractSetupCommand
     }
 
     /**
-     * @param InputInterface  $input
+     * @param InputInterface $input
      * @param OutputInterface $output
      *
      * @return void
@@ -39,12 +40,13 @@ class RemoveCommand extends AbstractSetupCommand
         $this->detectMagento($output, true);
 
         if ($this->initMagento()) {
-            $moduleName     = $this->getModule($input);
-            $setupName      = $input->getArgument('setup');
-            $moduleSetups   = $this->getModuleSetupResources($moduleName);
+            $moduleName = $this->getModule($input);
+            $setupName = $input->getArgument('setup');
+            $moduleSetups = $this->getModuleSetupResources($moduleName);
 
             if (empty($moduleSetups)) {
                 $output->writeln(sprintf('No setup resources found for module: "%s"', $moduleName));
+
                 return;
             }
 
@@ -68,9 +70,13 @@ class RemoveCommand extends AbstractSetupCommand
      */
     public function removeSetupResource($moduleName, $setupResource, OutputInterface $output)
     {
-        $model          = $this->_getModel('core/resource', 'Mage_Core_Model_Resource');
-        $table          = $model->getTableName('core_resource');
-        $writeAdapter   = $model->getConnection('core_write');
+        /** @var Mage_Core_Model_Resource $model */
+        $model = $this->_getModel('core/resource', 'Mage_Core_Model_Resource');
+        $writeAdapter = $model->getConnection('core_write');
+        if (!$writeAdapter) {
+            throw new RuntimeException('Database not configured');
+        }
+        $table = $model->getTableName('core_resource');
 
         if ($writeAdapter->delete($table, array('code = ?' => $setupResource)) > 0) {
             $output->writeln(